requirements.txt
matplotlib==3.1.1
svgwrite==1.3.1
Keras==2.2.4
fontawesome==5.7.2.post1
opencv_python==4.1.0.25
numpy==1.16.4
requests==2.21.0
pandas==0.24.2
tensorflow==1.15.2
pytz==2019.2
CairoSVG==2.4.0
astral==1.10.1
scikit_learn==0.22.2.post1
PyYAML==5.3.1
